Add Help Message with Usage Information
Add the option to run gcommit -h or --h to get a help message explaining
expected usage of GCommit.

Refactor argument parsing to use python's built-in argparse rather than
accessing sys.argv directly. This is necessary because the -h or --help
options would be read as part of sys.argv, so the previous list slicing
syntax broke.

Signed-off-by: "Jay Welborn <jay@jaywelborn.com>"